Fix/h2

+++ src/main/java/com/example/Tzms/service/ProductService.java
@@ -6,10 +6,14 @@ import com.example.Tzms.exception.ProductNotFoundException;
 import com.example.Tzms.model.Product;
 import com.example.Tzms.repository.ProductRepo;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 /**
@@ -64,8 +68,10 @@ public class ProductService {
      * Метод для получения всех товаров из бд
      * @return
      */
-    public List<ProductResponse> getAllProducts() {
-        List<Product> products = productRepo.findAll();
+    public List<ProductResponse> getAllProducts(Integer page) {
+        int pageNo = page < 1 ? 0 : page - 1;
+        Pageable pageable = PageRequest.of(pageNo, 10, Sort.Direction.DESC, "createdAt");
+        List<Product> products = productRepo.findAll(pageable).getContent();
         return products.stream().map(this::mapToProductResponse).toList();
     }
 
@@ -85,8 +91,14 @@ public class ProductService {
      * Метод для удаления товара из системы
      * @param id
      */
-    public void deleteProductByUuid(UUID id) {
-        productRepo.deleteById(id);
+    public void deleteProductByUuid(UUID id) throws ProductNotFoundException {
+        Optional<Product> productToRemove = productRepo.findById(id);
+
+        if (productToRemove.isPresent()) {
+            productRepo.deleteById(id);
+        } else {
+            throw new ProductNotFoundException("Товар с артикулом " + id + " не найден");
+        }
     }
 
     /**
